Header verification.
This version just reads and verifies the header of the index format v5, without
giving an error. If there is a correct index file of version v5 at .git/index
the output of git status -s is a list of all the files in the repository,
once as deleted and then with as unknown to git.
(Eg:
D  test
?? test )
The output of git ls-files is empty.