Reading the directories from index format v5.
All directories, with their data are read from the index format v5. To "test"
this, the git ls-files command is abused again. Since there are no files read
yet, the command doesn't output anything, but instead all directories with the
data are written to stdout. (For more explanation of the fields see
https://github.com/tgummerer/git/wiki/Index-file-format-v5). This is a very
loose test, since there is nothing to compare it with, but it gives good
indication if the code works.