Reading entries from index v5.
All index entries can now be read from index-v5. The entries are not
yet correctly mapped to the current in memory format. To "test" this
version, it outputs a list of all files, simulating the git ls-files
output. As for the directories, this is somewhat cheaty, but it works
well in practice to test the functionality.