apply: only pass required data to git_header_name
Currently the git_header_name function takes 'struct apply_state' as
parameter, even though it only needs the p_value from that struct.

This function is in the callchain of 'parse_git_header()', which we
want to make more generally useful in a subsequent commit.  To make
that happen we only want to pass in the required data to
'parse_git_header()', and not the whole 'struct apply_state'.

Only passing the required data to the functions in the callchain of
'parse_git_header()' will make that easier.

Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>