apply: only pass required data to gitdiff_* functions
Currently the gitdiff_* functions take 'struct apply_state' as
parameter, even though they only needs the root, linenr and p_value
from that struct.

These functions are in the callchain of 'parse_git_header()', which we
want to make more generally useful in a subsequent commit.  To make
that happen we only want to pass in the required data to
'parse_git_header()', and not the whole 'struct apply_state'.

Only passing the required data to the functions in the callchain of
'parse_git_header()' will make that easier.

Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>