treewide: format unsigned long using %lu
Currently we use %ld in some places to format unsigned longs. This may
overflow and print negative numbers in some cases which can result in
confusing output for the end user.

Change the formatting strings to use %lu for unsigned longss
everywhere to avoid that.

Some places converted here might not actually realistically overflow.
However it's easier to convert everything at once instead of checking
whether or not ever callsite can or can't overflow.  More importantly,
this brings us one step closer to turning on '-Wformat-signedness',
which will avoid similar problems once and for all in the future.

Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>