treewide: use unsigned numbers for %o
%o doesn't handle negative numbers, but we currently use it to format
unsigned ints.  Avoid that by casting the numbers when we know they
are positive, or passin unsigned constants as appropriate.

This casting/passing unsigned numbers isn't strictly necessasy in any
of the places this patch changes, however it stops the compiler from
complaining when -Wformat-signedness is passed, so it helps turning it
on in a further step.

Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>