archive-tar: use unsigned variables for unsigned values
The len and tmp variables can never be negative.  In fact we already
use %u to format len later in the function.  Make the variables
unsigned to match the expectations later in the function.

I don't think tmp or len would realistically overflow here, but this
came up trying to make the codebase compile with
"-Wformat-signedness", and it seems better to change the variable type
to fit the values better than to change the format string to match the
existing variable type.

Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>