treewide: specify field width as int
Format strings expect the field width (if given) to be specified by an
int.  The compiler will complain when compiling with
"-Wformat-signedness" otherwise.  Cast values that are not ints to int
when using them to specify the field length.

This is safe to do.  The formatting functions already expect an int
here, and if there is an overflow, this is a bug anyway which we
should fix separately.

Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>