show-index: make i unsigned
The variable i is used as loop counter, where it counts up to nr,
which is also unsigned.  i can never be negative throughout the
execution of show-index.

Make the variable itself unsigned to reflect that.

This was found while trying to make the codebase compile with
"-Wformat-signedness", as we already use %u to format i in some cases
throughout the function.

Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>