diff: make mode a unsigned short in prep_temp_blob
Make the mode argument in prep_temp_blob a short.  All callsites pass
a unsigned short to this function.  Nothing is theoretically wrong
with using an int, as it is the wider type.  However later in the
function we format this parameter using %o, which requires an unsigned
value.

Changing the type here will allow us to turn on "-Wformat-signedness"
in a later step.

Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>